fix(spin): make jsonPath use Double nodes as default number type
Related to https://github.com/camunda/camunda-bpm-platform/issues/2575

+++ spin/dataformat-json-jackson/src/test/java/org/camunda/spin/json/mapping/OrderDetails.java
@@ -22,7 +22,7 @@ public class OrderDetails {
 
   private String article;
   private double price;
-  private int roundedPrice;
+  private long roundedPrice;
   private List<String> currencies;
   private boolean paid;
 
@@ -38,10 +38,10 @@ public class OrderDetails {
   public void setPrice(double price) {
     this.price = price;
   }
-  public int getRoundedPrice() {
+  public long getRoundedPrice() {
     return roundedPrice;
   }
-  public void setRoundedPrice(int roundedPrice) {
+  public void setRoundedPrice(long roundedPrice) {
     this.roundedPrice = roundedPrice;
   }
   public List<String> getCurrencies() {
